refactor: BrokeLocation is not an enum
